+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.7.0-rc.44",
+  sdkVersion: "4.7.0-rc.45",
   genVersion: "2.879.6",
-  userAgent: "speakeasy-sdk/typescript 4.7.0-rc.44 2.879.6 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.7.0-rc.45 2.879.6 2.0 @orq-ai/node",
 } as const;